% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterCornerAxes.R
\name{clusterCornerAxes}
\alias{clusterCornerAxes}
\title{Add corner axes on seurat UMAP/tSNE cluster figures}
\arguments{
\item{object}{seurat object.}

\item{reduction}{"string", reduction type (umap/tsne).}

\item{group_facet}{"string", give the column name in seurat metadata to facet plot.}

\item{cluster_col}{"string", the point color to group by,cluster name, default "seurat_clusters".}

\item{p_size}{"num", point size.}

\item{aspect_ratio}{"num", plot width and height ratio, default NULL.}

\item{no_split}{"logic", whether to split/facet the plot, default "TRUE".}

\item{nrow}{"num", rows to plot when no_split = FALSE.}

\item{rel_length}{"num", the corner axis line relative length to plot axis(0-1).}

\item{rel_dist}{"num" ,the relative distance of corner axis label to axis.}

\item{axes}{"string", show multiple corner axis or only one (mul/one), default "mul".}

\item{legend_pos}{"string", legend position same as ggplot theme function, default "right".}

\item{key_size}{The legned point size, default is 5.}

\item{line_text_col}{"string", corner line and label color, default "black".}

\item{strip_col}{"string", facet background color, default "white".}

\item{arrow_type}{"string", arrow type (open/closed), default "closed".}

\item{corner_text_size}{"num", the corner label text size, default is 3.}

\item{base_size}{"num", theme base size, default is 14.}

\item{theme_bg}{Another theme style, default is "default", or "bwCorner".}

\item{add_circle}{"logic", whether add circle on clusters, default is "FALSE".}

\item{add_circle_legacy}{"logic", using the legacy version to add a circle, the parameters `nbin`, `nsm`, `addsm`, `sfac` and `qval` are only applicable to legacy, default is "FALSE".}

\item{cic_alpha}{"num", circle fill color alpha, default is 0.1.}

\item{cic_delta}{"num", the distance to extend the curve (circle), this parameter only takes effect when `add_circle_legacy = FALSE`.}

\item{cic_line_size}{"num", circle line size, default is 1.}

\item{cic_line_color}{"num", circle line color, default is "grey50".}

\item{cic_line_lty}{"num", circle line type, default is "dashed".}

\item{nbin}{"num", number of points used to shape the hull, default 100.}

\item{nsm}{"num", number of points used to perform convolution, should less than nbin, default 10.}

\item{addsm}{"num", number of additional times of convolution performed, default 1.}

\item{qval}{"num", expansion size factor, larger value means bigger hull, default 1.5.}

\item{sfac}{"num", quantile of each sector, used to determine the edge of the hull, should less than 1, default 1.5.}

\item{cell_label}{Whether to label cell type on plot, default is FALSE.}

\item{cell_label_size}{Cell type label size, default is 6.}

\item{cell_label_color}{Cell type label color, default is "black".}

\item{show_legend}{Whether show legend, default is TRUE.}

\item{corner_variable}{Which group corner axis to be added when "axes" set to "one", default is the first level.}
}
\value{
Return a ggplot object.
}
\description{
Add corner axes on seurat UMAP/tSNE cluster figures
}
\examples{
\dontrun{test <- system.file("extdata", "seuratTest.RDS", package = "scVisual")

tmp <- readRDS(test)

# umap
clusterCornerAxes(
  object = tmp, reduction = "umap",
  no_split = TRUE
)

# arrowType
clusterCornerAxes(
  object = tmp, reduction = "umap",
  no_split = TRUE, arrow_type = "open"
)

# facet by metadata column "orig.ident"
clusterCornerAxes(
  object = tmp, reduction = "umap",
  no_split = FALSE, group_facet = "orig.ident",
  rel_length = 0.5
)

# retain only one axes with factor ordering
tmp$orig.ident <- factor(tmp$orig.ident, levels = c("ST2", "ST3", "ST1", "ST4"))
clusterCornerAxes(
  object = tmp, reduction = "umap",
  no_split = FALSE, group_facet = "orig.ident",
  rel_length = 0.5,
  axes = "one"
)

# line color
clusterCornerAxes(
  object = tmp, reduction = "umap",
  no_split = FALSE, group_facet = "orig.ident",
  rel_length = 0.5,
  line_text_col = "grey50"
)

# tsne
clusterCornerAxes(
  object = tmp, reduction = "tsne",
  no_split = FALSE, group_facet = "orig.ident",
  rel_length = 0.5
)
}
}
\author{
mixfruit
}
