% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCornerAxes.R
\name{featureCornerAxes}
\alias{featureCornerAxes}
\title{Add corner axes on seurat UMAP/tSNE gene FeaturePlot function figures}
\arguments{
\item{object}{object seurat object.}

\item{reduction}{"string", reduction type (umap/tsne).}

\item{features}{"string", the gene you want to plot.}

\item{group_facet}{"string", give the column name in seurat metadata to facet plot, if it is "NULL", facet plot only by gene.}

\item{rel_length}{"num", the corner axis line relative length to plot axis(0-1).}

\item{rel_dist}{"num", the relative distance of corner axis label to axis.}

\item{aspect_ratio}{"num", plot width and height ratio, default NULL.}

\item{low}{"string", point color with low expression.}

\item{high}{"string", point color with high expression.}

\item{axes}{"string", show multiple corner axis or only one (mul/one), default "mul".}

\item{legend_pos}{"string", legend position same as ggplot theme function, default "right".}

\item{strip_col}{"string", facet background color, defaults "white".}

\item{p_size}{"num", point size.}

\item{arrow_type}{"string", arrow type (open/closed), default "closed".}

\item{line_text_col}{"string", facet background color, default "white".}

\item{corner_text_size}{"num", the corner label text size, default is 5.}

\item{base_size}{"num", theme base size, default is 14.}

\item{theme_bg}{Another theme style, default is "default", or "bwCorner".}

\item{show.legend}{Whether show legend, default is "TRUE".}

\item{cornerVariable}{Which group corner axis to be added when "axes" set to "one", default is the first group.}

\item{n_layout}{= NULL Similar to the ncol/nrow for the layout, default is the gene numbers.}

\item{min_exp}{Minimum expression value defined, default is NULL.}

\item{max_exp}{Maxmum expression value defined, default is NULL.}
}
\value{
Return a ggplot.
}
\description{
Add corner axes on seurat UMAP/tSNE gene FeaturePlot function figures
}
\examples{
\dontrun{

test <- system.file("extdata", "seuratTest.RDS", package = "scVisual")

tmp <- readRDS(test)

# umap
featureCornerAxes(
  object = tmp, reduction = "umap",
  group_facet = "orig.ident",
  rel_length = 0.5, rel_dist = 0.2,
  features = c("Actb", "Ythdc1", "Ythdf2")
)

# one axes with factor ordering
tmp$orig.ident <- factor(tmp$orig.ident, levels = c("ST2", "ST3", "ST1", "ST4"))
featureCornerAxes(
  object = tmp, reduction = "umap",
  group_facet = "orig.ident",
  features = c("Actb", "Ythdc1", "Ythdf2"),
  rel_length = 0.5, rel_dist = 0.2,
  axes = "one",
  line_text_col = "grey50"
)

# tsne
featureCornerAxes(
  object = tmp, reduction = "tsne",
  group_facet = "orig.ident",
  rel_length = 0.5, rel_dist = 0.2,
  features = c("Actb", "Ythdc1", "Ythdf2")
)

}
}
\author{
mixfruit
}
