% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-grid.R
\name{grid.colorkey}
\alias{grid.colorkey}
\title{Custom color key for grid plots}
\usage{
grid.colorkey(
  x = NULL,
  color = NULL,
  color.n = 100,
  ticks.side = c("left", "right", "top", "bottom"),
  pos = c("h", "v")
)
}
\arguments{
\item{x}{Data values for color scaling}

\item{color}{Color palette}

\item{color.n}{Number of color levels}

\item{ticks.side}{Side for ticks}

\item{pos}{Orientation (horizontal or vertical)}
}
\value{
None
}
\description{
Create a custom color key for grid plots
}
\author{
mixfruit
}
