% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-grid.R
\name{grid.yaxis2}
\alias{grid.yaxis2}
\title{Custom y-axis for grid plots}
\usage{
grid.yaxis2(
  at = NULL,
  breaks = 5,
  labels = NULL,
  tick.len = 0.5,
  label.space = 0.25,
  side = c("left", "right"),
  rot = 0,
  label.size = 12
)
}
\arguments{
\item{at}{Position of ticks}

\item{breaks}{Number of breaks}

\item{labels}{Tick labels}

\item{tick.len}{Tick length}

\item{label.space}{Space between ticks and labels}

\item{side}{Side of the axis (left or right)}

\item{rot}{Label rotation}

\item{label.size}{Label font size}
}
\value{
None
}
\description{
Create a custom y-axis for grid plots
}
\author{
mixfruit
}
