% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jjDotPlot.R
\name{jjDotPlot}
\alias{jjDotPlot}
\title{Dot Plot for Gene Expression Visualization}
\arguments{
\item{object}{Seurat object containing the data.}

\item{gene}{Genes to plot.}

\item{markerGene}{Marker genes with cell type info.}

\item{...}{Additional parameters.

Create a dot plot of average expression and percent expressing
across clusters or groups.}
}
\value{
A ggplot object.
}
\description{
Dot Plot for Gene Expression Visualization
}
\author{
mixfruit
}
