% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jjVolcano.R
\name{jjVolcano}
\alias{jjVolcano}
\title{using jjVolcano to visualize marker genes}
\arguments{
\item{diff_data}{diff results with data.frame format, default NULL.}

\item{my_markers}{whether supply your own gene labels, default NULL.}

\item{log2fc_cutoff}{log2FoldChange cutoff, default 0.25.}

\item{pvalue_cutoff}{pvalue cutoff to filter, default 0.05.}

\item{adjust_p_cutoff}{adjusted pvalue cutoff to be colored in plot, default 0.01.}

\item{top_gene_n}{top genes to be labeled in plot, default 5.}

\item{col_type}{point color type("updown/adjustP"), default "updown".}

\item{back_col}{background color, default "grey93".}

\item{p_size}{point size, default 0.75.}

\item{aes_col}{point mapping color, default c("#0099CC","#CC3333").}

\item{legend_position}{legend position in plot, default c(0.7,0.9).}

\item{base_size}{theme base size, default 14.}

\item{tile_col}{cluster tile fill color, default useMyCol("paired",n = 9).}

\item{...}{other arguments passed by "geom_text_repel".}

\item{cluster_order}{whether given your cluster orders, default NULL.}

\item{polar}{whether make the plot to br polar, default FALSE.}

\item{expand}{the y axis expand, default c(-1,1).}

\item{flip}{whether flip the plot, default FALSE.}

\item{order_by}{top marker gene selection method, how the order is, default c("avg_log2FC").}
}
\value{
a ggplot object.
}
\description{
using jjVolcano to visualize marker genes
}
\examples{
\dontrun{
jjVolcano(diff_data = pbmc.markers)
}
}
\author{
mixfruit
}
