% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerVolcano.R
\name{marker_volcano}
\alias{marker_volcano}
\alias{markerVolcano}
\title{Marker genes volcano plot}
\arguments{
\item{markers}{Dataframe marker genes from findAllmarkers function from seurat.}

\item{own_gene}{Your own gene names to be labeled on plot, defaults is null.}

\item{top_n}{Numbers top genes to label, defaults is 5.}

\item{log2fc}{The threshold of log2FC, defaults is 0.25.}

\item{hline_size}{Hline size, defaults is 1.}

\item{hline_color}{Hline color, defaults is "grey50".}

\item{p_force}{Positive gene force parameters to avoid overlap gene labels,
defaults is 5.}

\item{n_force}{Negative gene force parameters to avoid overlap gene labels,
defaults is 2.5.}

\item{nudge_x}{Adjustments on the horizontal of the gene label, defaults is 0.8.}

\item{p_nudge_y}{Adjustments on the horizontal of the positive gene label,
defaults is 0.25.}

\item{n_nudge_y}{Adjustments on the horizontal of the negative gene label,
defaults is 0.}

\item{base_size}{Theme base size, defaults is 14.}

\item{facet_color}{Facet border color, defaults is NA.}

\item{facet_fill}{Facet fill color, defaults is "white".}

\item{ylab}{Plot y label, defaults is "Log2-Fold Change".}

\item{nrow}{Numbers rows to plot, defaults is 1.}
}
\value{
Return a ggplot.
}
\description{
Marker genes volcano plot
}
\examples{
test <- system.file("extdata", "pbmc.markers.csv", package = "scVisual")
markers <- read.csv(test)

marker_volcano(
  markers = markers,
  top_n = 5,
  label_col = ggsci::pal_npg()(9)
)
}
\author{
mixfruit
}
