% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterCellPlot.R
\name{scatter_cell_plot}
\alias{scatter_cell_plot}
\alias{scatterCellPlot}
\title{Create a scatter cell plot using the grid package}
\usage{
scatter_cell_plot(
  object = NULL,
  color = NULL,
  dim = "umap",
  rm_axis = FALSE,
  cell_id = NULL,
  bar_width = 0.2,
  point_size = 1,
  rm_barplot = FALSE,
  legend_psize = 1.5,
  arrow_len = 0.2
)
}
\arguments{
\item{object}{A Seurat object containing the data.}

\item{color}{A vector of colors for each cell type. If NULL, random colors
will be assigned.}

\item{dim}{The dimension used for plotting (default is "umap").}

\item{rm_axis}{Logical value indicating whether to remove the x and y-axis
(default is FALSE).}

\item{cell_id}{Name of the column in the metadata that represents cell
identity (default is NULL).}

\item{bar_width}{Width of the barplot (default is 0.2).}

\item{point_size}{Size of the points in the scatter plot (default is 1).}

\item{rm_barplot}{Whether to remove barplot, default FALSE.}

\item{legend_psize}{Legend point size, default 1.5.}

\item{arrow_len}{Arrow length, default 0.2.}
}
\value{
None
}
\description{
Create a scatter cell plot using the grid package
}
\examples{
\dontrun{
scatter_cell_plot(object = seurat_object)
}

}
\author{
mixfruit
}
