% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-color.R
\name{useMyCol}
\alias{useMyCol}
\title{Custom color palette function}
\usage{
useMyCol(palette = "paired", n)
}
\arguments{
\item{palette}{Palette name}

\item{n}{Number of colors to generate}
}
\value{
A vector of colors
}
\description{
Generate a custom color palette
}
\author{
mixfruit
}
